refactor: 유저 관련 DTO 내부 클래스 분리
