+++ src/main/java/com/recipe/app/src/user/application/dto/UserRecipeResponse.java
@@ -1,0 +1,22 @@
+package com.recipe.app.src.user.application.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Builder;
+import lombok.Getter;
+
+@Schema(description = "나만의 레시피 응답 DTO")
+@Getter
+public class UserRecipeResponse {
+
+    @Schema(description = "레시피 고유 번호")
+    private final Long recipeId;
+    @Schema(description = "썸네일 이미지")
+    private final String thumbnailImgUrl;
+
+    @Builder
+    public UserRecipeResponse(Long recipeId, String thumbnailImgUrl) {
+
+        this.recipeId = recipeId;
+        this.thumbnailImgUrl = thumbnailImgUrl;
+    }
+}
